+++ lib/compat/wordpress-6.5/interactivity-api/class-wp-interactivity-api.php
@@ -141,20 +141,36 @@ if ( ! class_exists( 'WP_Interactivity_API' ) ) {
 		 * @since 6.5.0
 		 */
 		public function print_client_interactivity_data() {
-			$store      = array();
-			$has_state  = ! empty( $this->state_data );
-			$has_config = ! empty( $this->config_data );
+			if ( empty( $this->state_data ) && empty( $this->config_data ) ) {
+				return;
+			}
 
-			if ( $has_state || $has_config ) {
-				if ( $has_config ) {
-					$store['config'] = $this->config_data;
+			$interactivity_data = array();
+
+			$config = array();
+			foreach ( $this->config_data as $key => $value ) {
+				if ( ! empty( $value ) ) {
+					$config[ $key ] = $value;
 				}
-				if ( $has_state ) {
-					$store['state'] = $this->state_data;
+			}
+			if ( ! empty( $config ) ) {
+				$interactivity_data['config'] = $config;
+			}
+
+			$state = array();
+			foreach ( $this->state_data as $key => $value ) {
+				if ( ! empty( $value ) ) {
+					$state[ $key ] = $value;
 				}
+			}
+			if ( ! empty( $state ) ) {
+				$interactivity_data['state'] = $state;
+			}
+
+			if ( ! empty( $interactivity_data ) ) {
 				wp_print_inline_script_tag(
 					wp_json_encode(
-						$store,
+						$interactivity_data,
 						JSON_HEX_TAG | JSON_HEX_AMP
 					),
 					array(
@@ -205,6 +221,9 @@ if ( ! class_exists( 'WP_Interactivity_API' ) ) {
 		 * @return string The processed HTML content. It returns the original content when the HTML contains unbalanced tags.
 		 */
 		public function process_directives( string $html ): string {
+			if ( ! str_contains( $html, 'data-wp-' ) ) {
+				return $html;
+			}
 			$context_stack   = array();
 			$namespace_stack = array();
 			$result          = $this->process_directives_args( $html, $context_stack, $namespace_stack );
@@ -286,34 +305,41 @@ if ( ! class_exists( 'WP_Interactivity_API' ) ) {
 					}
 				}
 				/*
-				* If the matching opener tag didn't have any directives, it can skip the
-				* processing.
-				*/
+				 * If the matching opener tag didn't have any directives, it can skip the
+				 * processing.
+				 */
 				if ( 0 === count( $directives_prefixes ) ) {
 					continue;
 				}
 
-				/*
-				 * Sorts the attributes by the order of the `directives_processor` array
-				 * and checks what directives are present in this element. The processing
-				 * order is reversed for tag closers.
-				 */
-				$directives_prefixes = array_intersect(
-					$p->is_tag_closer()
-					? $directive_processor_prefixes_reversed
-					: $directive_processor_prefixes,
-					$directives_prefixes
+				// Directive processing might be different depending on if it is entering the tag or exiting it.
+				$modes = array(
+					'enter' => ! $p->is_tag_closer(),
+					'exit'  => $p->is_tag_closer() || ! $p->has_and_visits_its_closer_tag(),
 				);
+				foreach ( $modes as $mode => $should_run ) {
+					if ( ! $should_run ) {
+						continue;
+					}
 
-				// Executes the directive processors present in this element.
-				foreach ( $directives_prefixes as $directive_prefix ) {
-					$func = is_array( self::$directive_processors[ $directive_prefix ] )
-						? self::$directive_processors[ $directive_prefix ]
-						: array( $this, self::$directive_processors[ $directive_prefix ] );
-					call_user_func_array(
-						$func,
-						array( $p, &$context_stack, &$namespace_stack, &$tag_stack )
+					/*
+					 * Sorts the attributes by the order of the `directives_processor` array
+					 * and checks what directives are present in this element.
+					 */
+					$existing_directives_prefixes = array_intersect(
+						'enter' === $mode ? $directive_processor_prefixes : $directive_processor_prefixes_reversed,
+						$directives_prefixes
 					);
+					foreach ( $existing_directives_prefixes as $directive_prefix ) {
+						$func = is_array( self::$directive_processors[ $directive_prefix ] )
+							? self::$directive_processors[ $directive_prefix ]
+							: array( $this, self::$directive_processors[ $directive_prefix ] );
+
+						call_user_func_array(
+							$func,
+							array( $p, $mode, &$context_stack, &$namespace_stack, &$tag_stack )
+						);
+					}
 				}
 			}
 
@@ -466,12 +492,13 @@ if ( ! class_exists( 'WP_Interactivity_API' ) ) {
 		 * @since 6.5.0
 		 *
 		 * @param WP_Interactivity_API_Directives_Processor $p               The directives processor instance.
+		 * @param string                                    $mode            Whether the processing is entering or exiting the tag.
 		 * @param array                                     $context_stack   The reference to the context stack.
 		 * @param array                                     $namespace_stack The reference to the store namespace stack.
 		 */
-		private function data_wp_interactive_processor( WP_Interactivity_API_Directives_Processor $p, array &$context_stack, array &$namespace_stack ) {
-			// In closing tags, it removes the last namespace from the stack.
-			if ( $p->is_tag_closer() ) {
+		private function data_wp_interactive_processor( WP_Interactivity_API_Directives_Processor $p, string $mode, array &$context_stack, array &$namespace_stack ) {
+			// When exiting tags, it removes the last namespace from the stack.
+			if ( 'exit' === $mode ) {
 				array_pop( $namespace_stack );
 				return;
 			}
@@ -510,12 +537,13 @@ if ( ! class_exists( 'WP_Interactivity_API' ) ) {
 		 * @since 6.5.0
 		 *
 		 * @param WP_Interactivity_API_Directives_Processor $p               The directives processor instance.
+		 * @param string                                    $mode            Whether the processing is entering or exiting the tag.
 		 * @param array                                     $context_stack   The reference to the context stack.
 		 * @param array                                     $namespace_stack The reference to the store namespace stack.
 		 */
-		private function data_wp_context_processor( WP_Interactivity_API_Directives_Processor $p, array &$context_stack, array &$namespace_stack ) {
-			// In closing tags, it removes the last context from the stack.
-			if ( $p->is_tag_closer() ) {
+		private function data_wp_context_processor( WP_Interactivity_API_Directives_Processor $p, string $mode, array &$context_stack, array &$namespace_stack ) {
+			// When exiting tags, it removes the last context from the stack.
+			if ( 'exit' === $mode ) {
 				array_pop( $context_stack );
 				return;
 			}
@@ -556,11 +584,12 @@ if ( ! class_exists( 'WP_Interactivity_API' ) ) {
 		 * @since 6.5.0
 		 *
 		 * @param WP_Interactivity_API_Directives_Processor $p               The directives processor instance.
+		 * @param string                                    $mode            Whether the processing is entering or exiting the tag.
 		 * @param array                                     $context_stack   The reference to the context stack.
 		 * @param array                                     $namespace_stack The reference to the store namespace stack.
 		 */
-		private function data_wp_bind_processor( WP_Interactivity_API_Directives_Processor $p, array &$context_stack, array &$namespace_stack ) {
-			if ( ! $p->is_tag_closer() ) {
+		private function data_wp_bind_processor( WP_Interactivity_API_Directives_Processor $p, string $mode, array &$context_stack, array &$namespace_stack ) {
+			if ( 'enter' === $mode ) {
 				$all_bind_directives = $p->get_attribute_names_with_prefix( 'data-wp-bind--' );
 
 				foreach ( $all_bind_directives as $attribute_name ) {
@@ -572,7 +601,10 @@ if ( ! class_exists( 'WP_Interactivity_API' ) ) {
 					$attribute_value = $p->get_attribute( $attribute_name );
 					$result          = $this->evaluate( $attribute_value, end( $namespace_stack ), end( $context_stack ) );
 
-					if ( null !== $result && ( false !== $result || '-' === $bound_attribute[4] ) ) {
+					if ( null !== $result && (
+						false !== $result ||
+						( strlen( $bound_attribute ) > 5 && '-' === $bound_attribute[4] )
+					) ) {
 						/*
 						 * If the result of the evaluation is a boolean and the attribute is
 						 * `aria-` or `data-, convert it to a string "true" or "false". It
@@ -580,7 +612,10 @@ if ( ! class_exists( 'WP_Interactivity_API' ) ) {
 						 * replicate what Preact will later do in the client:
 						 * https://github.com/preactjs/preact/blob/ea49f7a0f9d1ff2c98c0bdd66aa0cbc583055246/src/diff/props.js#L131C24-L136
 						 */
-						if ( is_bool( $result ) && '-' === $bound_attribute[4] ) {
+						if (
+							is_bool( $result ) &&
+							( strlen( $bound_attribute ) > 5 && '-' === $bound_attribute[4] )
+						) {
 							$result = $result ? 'true' : 'false';
 						}
 						$p->set_attribute( $bound_attribute, $result );
@@ -600,11 +635,12 @@ if ( ! class_exists( 'WP_Interactivity_API' ) ) {
 		 * @since 6.5.0
 		 *
 		 * @param WP_Interactivity_API_Directives_Processor $p               The directives processor instance.
+		 * @param string                                    $mode            Whether the processing is entering or exiting the tag.
 		 * @param array                                     $context_stack   The reference to the context stack.
 		 * @param array                                     $namespace_stack The reference to the store namespace stack.
 		 */
-		private function data_wp_class_processor( WP_Interactivity_API_Directives_Processor $p, array &$context_stack, array &$namespace_stack ) {
-			if ( ! $p->is_tag_closer() ) {
+		private function data_wp_class_processor( WP_Interactivity_API_Directives_Processor $p, string $mode, array &$context_stack, array &$namespace_stack ) {
+			if ( 'enter' === $mode ) {
 				$all_class_directives = $p->get_attribute_names_with_prefix( 'data-wp-class--' );
 
 				foreach ( $all_class_directives as $attribute_name ) {
@@ -634,11 +670,12 @@ if ( ! class_exists( 'WP_Interactivity_API' ) ) {
 		 * @since 6.5.0
 		 *
 		 * @param WP_Interactivity_API_Directives_Processor $p               The directives processor instance.
+		 * @param string                                    $mode            Whether the processing is entering or exiting the tag.
 		 * @param array                                     $context_stack   The reference to the context stack.
 		 * @param array                                     $namespace_stack The reference to the store namespace stack.
 		 */
-		private function data_wp_style_processor( WP_Interactivity_API_Directives_Processor $p, array &$context_stack, array &$namespace_stack ) {
-			if ( ! $p->is_tag_closer() ) {
+		private function data_wp_style_processor( WP_Interactivity_API_Directives_Processor $p, string $mode, array &$context_stack, array &$namespace_stack ) {
+			if ( 'enter' === $mode ) {
 				$all_style_attributes = $p->get_attribute_names_with_prefix( 'data-wp-style--' );
 
 				foreach ( $all_style_attributes as $attribute_name ) {
@@ -726,11 +763,12 @@ if ( ! class_exists( 'WP_Interactivity_API' ) ) {
 		 * @since 6.5.0
 		 *
 		 * @param WP_Interactivity_API_Directives_Processor $p               The directives processor instance.
+		 * @param string                                    $mode            Whether the processing is entering or exiting the tag.
 		 * @param array                                     $context_stack   The reference to the context stack.
 		 * @param array                                     $namespace_stack The reference to the store namespace stack.
 		 */
-		private function data_wp_text_processor( WP_Interactivity_API_Directives_Processor $p, array &$context_stack, array &$namespace_stack ) {
-			if ( ! $p->is_tag_closer() ) {
+		private function data_wp_text_processor( WP_Interactivity_API_Directives_Processor $p, string $mode, array &$context_stack, array &$namespace_stack ) {
+			if ( 'enter' === $mode ) {
 				$attribute_value = $p->get_attribute( 'data-wp-text' );
 				$result          = $this->evaluate( $attribute_value, end( $namespace_stack ), end( $context_stack ) );
 
@@ -823,10 +861,11 @@ HTML;
 		 *
 		 * @since 6.5.0
 		 *
-		 * @param WP_Interactivity_API_Directives_Processor $p The directives processor instance.
+		 * @param WP_Interactivity_API_Directives_Processor $p               The directives processor instance.
+		 * @param string                                    $mode            Whether the processing is entering or exiting the tag.
 		 */
-		private function data_wp_router_region_processor( WP_Interactivity_API_Directives_Processor $p ) {
-			if ( ! $p->is_tag_closer() && ! $this->has_processed_router_region ) {
+		private function data_wp_router_region_processor( WP_Interactivity_API_Directives_Processor $p, string $mode ) {
+			if ( 'enter' === $mode && ! $this->has_processed_router_region ) {
 				$this->has_processed_router_region = true;
 
 				// Initialize the `core/router` store.
@@ -862,12 +901,13 @@ HTML;
 		 * @since 6.5.0
 		 *
 		 * @param WP_Interactivity_API_Directives_Processor $p               The directives processor instance.
+		 * @param string                                    $mode            Whether the processing is entering or exiting the tag.
 		 * @param array                                     $context_stack   The reference to the context stack.
 		 * @param array                                     $namespace_stack The reference to the store namespace stack.
 		 * @param array                                     $tag_stack       The reference to the tag stack.
 		 */
-		private function data_wp_each_processor( WP_Interactivity_API_Directives_Processor $p, array &$context_stack, array &$namespace_stack, array &$tag_stack ) {
-			if ( ! $p->is_tag_closer() && 'TEMPLATE' === $p->get_tag() ) {
+		private function data_wp_each_processor( WP_Interactivity_API_Directives_Processor $p, string $mode, array &$context_stack, array &$namespace_stack, array &$tag_stack ) {
+			if ( 'enter' === $mode && 'TEMPLATE' === $p->get_tag() ) {
 				$attribute_name   = $p->get_attribute_names_with_prefix( 'data-wp-each' )[0];
 				$extracted_suffix = $this->extract_prefix_and_suffix( $attribute_name );
 				$item_name        = isset( $extracted_suffix[1] ) ? $this->kebab_to_camel_case( $extracted_suffix[1] ) : 'item';