+++ packages/components/src/custom-select-control-v2/index.tsx
@@ -1,5 +1,4 @@
 /**
  * Internal dependencies
  */
-export { default as CustomSelect } from './default-component';
-export { default as CustomSelectItem } from './custom-select-item';
+export { default } from './default-component';