+++ packages/react-native-editor/CHANGELOG.md
@@ -10,6 +10,18 @@ For each user feature we should also add a importance categorization label  to i
 -->
 
 ## Unreleased
+
+## 1.117.0
+-   [*] Add empty fallback option for the BottomSheetSelectControl component [#60333]
+-   [*] Fix Quote Block styles [#60476]
+-   [*] Prevent passing potential false values to the onPress prop [#60595]
+-   [*] ColorPalette - Check for ScrollView reference [#60562]
+-   [*] Raw Handling - msListIgnore - Check attributes are valid [#60375]
+
+## 1.116.0
+-   [**] Highlight color formatting style improvements [#57650]
+
+## 1.115.0
 -   [*] Improve consistency of the block outline indicating the currently selected block [#59415]
 -   [*] Gallery block items with in-progress, paused, or failed media uploads correctly display an highlight outline when selected [#59423]
 -   [**] [internal] Upgrade React Native to version 0.73.3 [#58475]